separate a file into two files by class
